+++ be/src/vec/exec/format/table/transactional_hive_reader.h
@@ -17,24 +17,21 @@
 
 #pragma once
 
-#include <stddef.h>
-#include <stdint.h>
-
+#include <cstddef>
+#include <cstdint>
 #include <string>
-#include <tuple>
 #include <unordered_map>
 #include <unordered_set>
-#include <utility>
 #include <vector>
 
 #include "common/status.h"
 #include "exec/olap_common.h"
 #include "table_format_reader.h"
 #include "util/runtime_profile.h"
-#include "vec/columns/column_dictionary.h"
 #include "vec/common/hash_table/phmap_fwd_decl.h"
 
 namespace doris {
+#include "common/compile_check_begin.h"
 class RuntimeState;
 class SlotDescriptor;
 class TFileRangeDesc;
@@ -57,7 +54,7 @@ class TransactionalHiveReader : public TableFormatReader {
 public:
     struct AcidRowID {
         int64_t original_transaction;
-        int32_t bucket;
+        int64_t bucket;
         int64_t row_id;
 
         struct Hash {
@@ -137,4 +134,5 @@ inline bool operator<(const TransactionalHiveReader::AcidRowID& lhs,
 }
 
 } // namespace vectorized
+#include "common/compile_check_end.h"
 } // namespace doris
